refactor to allow for better testing of validators + step-4 isValid test
