+++ src/modules/create-market/validators/validate-initial-fair-prices.js
@@ -1,0 +1,34 @@
+export default function validateInitialFairPrices(type, initialFairPrices, width, halfWidth, scalarMin, scalarMax) {
+    // -- Constraints --
+    // 	Binary + Categorical:
+    //		min: priceWidth / 2
+    //  	max: 1 - (priceWidth / 2)
+    // 	Scalar:
+    // 		min: scalarMin + (priceWidth / 2)
+    // 		max: scalarMax - (priceWidth / 2)
+
+    const max = type === SCALAR ? parseFloat(scalarMax) - halfWidth : 1 - halfWidth;
+    const min = type === SCALAR ? parseFloat(scalarMin) + halfWidth : halfWidth;
+    const fairPriceErrors = {};
+
+    initialFairPrices.map((cV, i) => {
+        const parsed = parseFloat(cV);
+
+        if (!cV) {
+            fairPriceErrors[`${i}`] = 'Please provide some initial liquidity';
+            return;
+        }
+        if (Number.isNaN(parsed) && !Number.isFinite(parsed)) {
+            fairPriceErrors[`${i}`] = 'Initial liquidity must be numeric';
+            return;
+        }
+        if (cV < min || cV > max) {
+            fairPriceErrors[`${i}`] = `Initial prices must be between ${min} - ${max} based on the price width of ${width}`;
+            return;
+        }
+    });
+
+    if (!!Object.keys(fairPriceErrors).length) {
+        return fairPriceErrors;
+    }
+};
